requirements.txt
instaloader
python-telegram-bot
psycopg2